requirements.txt
python-dotenv
requests
Pillow
guidance
ipython
openai
markdownify
langchain_openai
gradio
langchain_experimental
fastapi
langchain_community
azure-ai-documentintelligence
uvicorn
llama-index-embeddings-azure-openai
llama-index
llama-index-llms-azure-openai
llama-index-embeddings-langchain
llama-index-llms-langchain